# documents/api-doc.yaml
openapi: 3.0.1
info:
  title: CRM - The Agile Monkeys
  version: v1
paths:
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    put:
      tags:
        - Users
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: request
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    delete:
      tags:
        - Users
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/users/{id}/recover:
    put:
      tags:
        - Users
      operationId: recoverUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/customers/{id}:
    get:
      tags:
        - Customers
      operationId: getCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    put:
      tags:
        - Customers
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    delete:
      tags:
        - Customers
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/customers/{id}/recover:
    put:
      tags:
        - Customers
      operationId: recoverCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/users:
    get:
      tags:
        - Users
      operationId: listUsers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    post:
      tags:
        - Users
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/customers:
    get:
      tags:
        - Customers
      operationId: listCustomers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    post:
      tags:
        - Customers
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /api/v1/customers/{id}/image:
    get:
      tags:
        - Customers
      operationId: getCustomerImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
components:
  schemas:
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
        admin:
          type: boolean
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        admin:
          type: boolean
        removed:
          type: boolean
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        surname2:
          type: string
        country:
          type: string
        document:
          type: string
        image64:
          type: string
    ChangeHistoryResponse:
      type: object
      properties:
        customerId:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
        description:
          type: string
        date:
          type: string
          format: date-time
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        surname:
          type: string
        surname2:
          type: string
        country:
          type: string
        document:
          type: string
        imageURL:
          type: string
        image64:
          type: string
        history:
          type: array
          items:
            $ref: '#/components/schemas/ChangeHistoryResponse'
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        admin:
          type: boolean
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        surname2:
          type: string
        country:
          type: string
        document:
          type: string
        image64:
          type: string